VideoRAGQnA/ui/requirements.txt
langchain-experimental
open-clip-torch
streamlit
sentence-transformers
